Reorganized structure a bit, trying to get singlepage interface working with server.
