+++ client/src/index.jsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+import App from '../src/App/app';
+
+ReactDOM.render(<App />, document.getElementById('app'));
